+++ packages/concerto-core/api.txt
@@ -7,7 +7,7 @@ class BaseFileException extends BaseException {
    + string getShortMessage() 
    + string getFileName() 
 }
-   + void checkConcertoObject() 
+   + void getTypeDeclaration() 
    + string getIdentifier() 
    + boolean isIdentifiable() 
    + boolean isRelationship()
