+++ packages/concerto-core/index.js
@@ -52,3 +52,17 @@ module.exports.TransactionDeclaration = require('./lib/introspect/transactiondec
 module.exports.Typed = require('./lib/model/typed');
 module.exports.ModelUtil = require('./lib/modelutil');
 module.exports.version = require('./package.json');
+
+module.exports.Concerto = {
+    getTypeDeclaration : require('./lib/concerto').getTypeDeclaration,
+    getIdentifier : require('./lib/concerto').getIdentifier,
+    setIdentifier : require('./lib/concerto').setIdentifier,
+    getFullyQualifiedIdentifier : require('./lib/concerto').getFullyQualifiedIdentifier,
+    toURI : require('./lib/concerto').toURI,
+    getType : require('./lib/concerto').getType,
+    getNamespace : require('./lib/concerto').getNamespace,
+    instanceOf : require('./lib/concerto').instanceOf,
+    isIdentifiable : require('./lib/concerto').isIdentifiable,
+    isRelationship : require('./lib/concerto').isRelationship,
+    validate : require('./lib/validator').validate,
+};
